split fileutils package into two new packages pkg/furl and pkg/filecopy
