+++ pkg/filecopy/mocks/Checker.go
