+++ pkg/furl/paths_test.go
@@ -14,36 +14,36 @@
  * limitations under the License.
  */
 
-package fileutils_test
+package furl_test
 
 import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/pivotal/go-ape"
+	"github.com/pivotal/go-ape/pkg/furl"
 )
 
 var _ = Describe("StartsWithHomeDirAsTilde", func() {
 
 	It("returns true when starting with ~/", func() {
-		result := fileutils.StartsWithCurrentUserDirectoryAsTilde("~/", "darwin")
+		result := furl.StartsWithCurrentUserDirectoryAsTilde("~/", "darwin")
 
 		Expect(result).To(BeTrue(), "tilde+forward slash should work")
 	})
 
 	It(`returns false when starting with ~\ on Mac OS`, func() {
-		result := fileutils.StartsWithCurrentUserDirectoryAsTilde(`~\`, "darwin")
+		result := furl.StartsWithCurrentUserDirectoryAsTilde(`~\`, "darwin")
 
 		Expect(result).To(BeFalse(), "tilde+backslash should not work on Mac OS")
 	})
 
 	It(`returns true when starting with ~\ on Windows`, func() {
-		result := fileutils.StartsWithCurrentUserDirectoryAsTilde(`~\`, "windows")
+		result := furl.StartsWithCurrentUserDirectoryAsTilde(`~\`, "windows")
 
 		Expect(result).To(BeTrue(), "tilde+backslash on Windows should work")
 	})
 
 	It(`returns true when starting with ~/ on Windows`, func() {
-		result := fileutils.StartsWithCurrentUserDirectoryAsTilde(`~/`, "windows")
+		result := furl.StartsWithCurrentUserDirectoryAsTilde(`~/`, "windows")
 
 		Expect(result).To(BeTrue(), "tilde+forward slash on Windows should work")
 	})
@@ -54,7 +54,7 @@ var _ = Describe("ResolveTilde", func() {
 	It("resolves ~/ against current user's home directory", func() {
 		initialPath := "~/some/location"
 
-		path, err := fileutils.ResolveTilde(initialPath)
+		path, err := furl.ResolveTilde(initialPath)
 
 		Expect(err).NotTo(HaveOccurred())
 		Expect(path).NotTo(ContainSubstring("~"))
@@ -64,7 +64,7 @@ var _ = Describe("ResolveTilde", func() {
 	It("returns path without tilde as is", func() {
 		initialPath := "look/matilde/no/tilde"
 
-		path, err := fileutils.ResolveTilde(initialPath)
+		path, err := furl.ResolveTilde(initialPath)
 
 		Expect(err).NotTo(HaveOccurred())
 		Expect(path).To(Equal(initialPath))
@@ -73,7 +73,7 @@ var _ = Describe("ResolveTilde", func() {
 	It("returns path with embedded tilde as is", func() {
 		initialPath := "look/matilde/thereisa/~"
 
-		path, err := fileutils.ResolveTilde(initialPath)
+		path, err := furl.ResolveTilde(initialPath)
 
 		Expect(err).NotTo(HaveOccurred())
 		Expect(path).To(Equal(initialPath))
